+++ ackpine-core/src/main/kotlin/ru/solrudev/ackpine/impl/installer/session/SessionBasedInstallSession.kt
@@ -42,12 +42,13 @@ import ru.solrudev.ackpine.impl.database.dao.SessionDao
 import ru.solrudev.ackpine.impl.database.dao.SessionFailureDao
 import ru.solrudev.ackpine.impl.database.dao.SessionProgressDao
 import ru.solrudev.ackpine.impl.installer.activity.SessionBasedInstallCommitActivity
-import ru.solrudev.ackpine.impl.installer.session.helpers.STREAM_COPY_PROGRESS_MAX
+import ru.solrudev.ackpine.impl.installer.session.helpers.PROGRESS_MAX
 import ru.solrudev.ackpine.impl.installer.session.helpers.copyTo
 import ru.solrudev.ackpine.impl.installer.session.helpers.openAssetFileDescriptor
 import ru.solrudev.ackpine.impl.session.AbstractProgressSession
 import ru.solrudev.ackpine.impl.session.helpers.CANCEL_CURRENT_FLAGS
 import ru.solrudev.ackpine.impl.session.helpers.commitSession
+import ru.solrudev.ackpine.impl.session.helpers.getSessionBasedSessionCommitProgressValue
 import ru.solrudev.ackpine.impl.session.helpers.launchConfirmation
 import ru.solrudev.ackpine.installer.InstallFailure
 import ru.solrudev.ackpine.installer.parameters.InstallMode
@@ -107,7 +108,8 @@ internal class SessionBasedInstallSession internal constructor(
 		if (initialState.isTerminal) {
 			return
 		}
-		if (initialProgress.progress >= 81) { // means that actual installation is ongoing or is completed
+		if (initialProgress.progress >= (context.getSessionBasedSessionCommitProgressValue() * PROGRESS_MAX).toInt()) {
+			// means that actual installation is ongoing or is completed
 			notifyCommitted() // block clients from committing
 		}
 		executor.executeWithSemaphore(nativeSessionIdSemaphore) {
@@ -209,7 +211,7 @@ internal class SessionBasedInstallSession internal constructor(
 		val future = ResolvableFuture.create<Unit>()
 		val countdown = AtomicInteger(apks.size)
 		val currentProgress = AtomicInteger(0)
-		val progressMax = apks.size * STREAM_COPY_PROGRESS_MAX
+		val progressMax = apks.size * PROGRESS_MAX
 		apks.forEachIndexed { index, uri ->
 			val afd = context.openAssetFileDescriptor(uri, cancellationSignal)
 				?: error("AssetFileDescriptor was null: $uri")
@@ -273,14 +275,14 @@ internal class SessionBasedInstallSession internal constructor(
 	}
 
 	private fun packageInstallerSessionCallback(nativeSessionId: Int) = object : PackageInstaller.SessionCallback() {
-		override fun onCreated(sessionId: Int) {}
-		override fun onBadgingChanged(sessionId: Int) {}
-		override fun onActiveChanged(sessionId: Int, active: Boolean) {}
-		override fun onFinished(sessionId: Int, success: Boolean) {}
+		override fun onCreated(sessionId: Int) { /* no-op */ }
+		override fun onBadgingChanged(sessionId: Int) { /* no-op */ }
+		override fun onActiveChanged(sessionId: Int, active: Boolean) { /* no-op */ }
+		override fun onFinished(sessionId: Int, success: Boolean) { /* no-op */ }
 
 		override fun onProgressChanged(sessionId: Int, progress: Float) {
 			if (sessionId == nativeSessionId) {
-				this@SessionBasedInstallSession.progress = Progress((progress * 100).toInt(), 100)
+				setProgress((progress * PROGRESS_MAX).toInt())
 			}
 		}
 	}
@@ -288,7 +290,7 @@ internal class SessionBasedInstallSession internal constructor(
 	private fun abandonSession() {
 		try {
 			packageInstaller.abandonSession(nativeSessionId)
-		} catch (_: Throwable) {
+		} catch (_: Throwable) { // no-op
 		}
 	}