+++ sample-java/src/main/java/ru/solrudev/ackpine/sample/uninstall/UninstallViewModel.java
@@ -51,8 +51,8 @@ import ru.solrudev.ackpine.uninstaller.parameters.UninstallParameters;
 
 public final class UninstallViewModel extends ViewModel {
 
-	private final static String SESSION_ID_KEY = "SESSION_ID";
-	private final static String PACKAGE_NAME_KEY = "PACKAGE_NAME";
+	private static final String SESSION_ID_KEY = "SESSION_ID";
+	private static final String PACKAGE_NAME_KEY = "PACKAGE_NAME";
 	private final MutableLiveData<Boolean> isLoading = new MutableLiveData<>(false);
 	private final MutableLiveData<List<ApplicationData>> applications = new MutableLiveData<>(new ArrayList<>());
 	private final DisposableSubscriptionContainer subscriptions = new DisposableSubscriptionContainer();
@@ -138,7 +138,7 @@ public final class UninstallViewModel extends ViewModel {
 			}
 
 			@Override
-			public void onFailure(@NonNull Throwable t) {
+			public void onFailure(@NonNull Throwable t) { // no-op
 			}
 		}, MoreExecutors.directExecutor());
 	}
@@ -153,7 +153,7 @@ public final class UninstallViewModel extends ViewModel {
 			}
 
 			@Override
-			public void onFailure(@NonNull Throwable t) {
+			public void onFailure(@NonNull Throwable t) { // no-op
 			}
 		}, MoreExecutors.directExecutor());
 	}