+++ sample-ktx/src/main/kotlin/ru/solrudev/ackpine/sample/install/SessionDataRepositoryImpl.kt
@@ -17,8 +17,8 @@
 package ru.solrudev.ackpine.sample.install
 
 import androidx.lifecycle.SavedStateHandle
+import ru.solrudev.ackpine.resources.ResolvableString
 import ru.solrudev.ackpine.session.Progress
-import ru.solrudev.ackpine.session.parameters.NotificationString
 import java.util.UUID
 
 private const val SESSIONS_KEY = "SESSIONS"
@@ -67,22 +67,23 @@ class SessionDataRepositoryImpl(private val savedStateHandle: SavedStateHandle) 
 			sessionsProgress[sessionProgressIndex] = SessionProgress(id, progress)
 		}
 		_sessionsProgress = sessionsProgress
-		if (progress.progress <= 80) {
-			return
-		}
-		val sessionDataIndex = _sessions.indexOfFirst { it.id == id }
-		if (sessionDataIndex != -1) {
-			val sessionData = _sessions[sessionDataIndex]
-			if (!sessionData.isCancellable) {
-				return
-			}
-			val sessions = _sessions.toMutableList()
-			sessions[sessionDataIndex] = sessionData.copy(isCancellable = false)
-			_sessions = sessions
-		}
 	}
 
-	override fun setError(id: UUID, error: NotificationString) {
+	override fun updateSessionIsCancellable(id: UUID, isCancellable: Boolean) {
+		val sessionDataIndex = _sessions.indexOfFirst { it.id == id }
+		if (sessionDataIndex == -1) {
+			return
+		}
+		val sessionData = _sessions[sessionDataIndex]
+		if (sessionData.isCancellable == isCancellable) {
+			return
+		}
+		val sessions = _sessions.toMutableList()
+		sessions[sessionDataIndex] = sessionData.copy(isCancellable = isCancellable)
+		_sessions = sessions
+	}
+
+	override fun setError(id: UUID, error: ResolvableString) {
 		val sessions = _sessions.toMutableList()
 		val sessionDataIndex = sessions.indexOfFirst { it.id == id }
 		if (sessionDataIndex != -1) {